requirements.txt
Flask-WTF
Flask-SQLAlchemy
Flask-Login